% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atpol_grids.R
\name{atpol100k}
\alias{atpol100k}
\title{atpol100k creates ATPOL grid 100km x 100km and returns it as sf object}
\usage{
atpol100k()
}
\value{
Simple Feature (sf) grid of polygons for 100km x 100km ATPOL grid
}
\description{
atpol100k creates ATPOL grid 100km x 100km and returns it as sf object
}
