% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atpol_grids.R
\name{atpol10k}
\alias{atpol10k}
\title{atpol10k returns ATPOL grid 10x10 km and returns it as sf object}
\usage{
atpol10k()
}
\value{
Simple Feature (sf) grid of polygons for 10km x 10km ATPOL grid
}
\description{
atpol10k returns ATPOL grid 10x10 km and returns it as sf object
}
