% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atpol_grids.R
\name{atpol1k}
\alias{atpol1k}
\title{atpol1k creates ATPOL grid 1km x 1km and returns it as sf object}
\usage{
atpol1k(grid)
}
\arguments{
\item{grid}{any valid ATPOL 10km grid like "BE23" or "DC58"}
}
\value{
Simple Feature (sf) grid of polygons for 1km x 1km ATPOL grid
}
\description{
atpol1k creates ATPOL grid 1km x 1km and returns it as sf object
}
\examples{
atpol1k("BE23")

}
