% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atpol_grids.R
\name{atpol_div_2}
\alias{atpol_div_2}
\title{atpol_div_2 creates ATPOL grid divided by 2 and returns it as sf object. Useful for grids like 5 x 5 km, 500 x 500 m or 50 x 50 m. For details see \insertCite{vereyStandaryzacjaZapisuPodzialow2018;textual}{atpolR}}
\usage{
atpol_div_2(grid)
}
\arguments{
\item{grid}{any valid ATPOL grid like "BE23" or "DC5128"}
}
\value{
Simple Feature (sf) grid of polygons for ATPOL grid divided by 2, ie. 5 km or 500 m, or 50 m
}
\description{
atpol_div_2 creates ATPOL grid divided by 2 and returns it as sf object. Useful for grids like 5 x 5 km, 500 x 500 m or 50 x 50 m. For details see \insertCite{vereyStandaryzacjaZapisuPodzialow2018;textual}{atpolR}
}
\examples{
atpol_div_2("BE23")
atpol_div_2(grid = c("BE23", "DC5128"))
}
\references{
\insertAllCited{}
}
