% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaryPL.R
\name{boundaryPL}
\alias{boundaryPL}
\title{boundaryPL reads the file data/pl_boundary.Rds with simplified boundary geometry.}
\usage{
boundaryPL()
}
\value{
Simple Feature (sf) geometry of Poland in EPSG:2180 projection.
}
\description{
boundaryPL reads the file data/pl_boundary.Rds with simplified boundary geometry.
}
