% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_atpol_square.R
\name{check_atpol_square}
\alias{check_atpol_square}
\title{Reverse engineering of published ATPOL grids}
\usage{
check_atpol_square(centroid, raster, distance)
}
\arguments{
\item{centroid}{Simple Feature point geometry for which the check is performed,
usually it corresponds to centroid of ATPOL 10km x 10km grid}

\item{raster}{geocoded raster, it has to be in EPSG:2180 projection}

\item{distance}{\code{st_buffer} distance from centroid point for which the check is done, default 1200 m}
}
\value{
"YES" or "?" for given SF point
}
\description{
check_atpol_square() do a reverse engineering of published ATPOL grids species,
especially those published in \insertCite{zajacAtlasRozmieszczeniaRoslin2001;textual}{atpolR}
}
\references{
\insertAllCited{}
}
