% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_atpol_square.R
\name{checkATPOLSquare}
\alias{checkATPOLSquare}
\title{Reverse engineering of published ATPOL grids}
\usage{
checkATPOLSquare(centroid, raster, distance)
}
\arguments{
\item{centroid}{SF point geometry for which the check is performed,
usually it corresponds to centroid of ATPOL 10km x 10km grid}

\item{raster}{geocoded raster, it has to be in EPSG:2180 projection}

\item{distance}{st_buffer distance from centroid point for which the check is done, default 1200 m}
}
\value{
"YES" or "?" for given SF point
}
\description{
checkATPOLSquare() do a reverse enginnering of published ATPOL grids species,
especially those published in: Zając, Adam, and Maria Zając, eds. 2001.
Atlas rozmieszczenia roślin naczyniowych w Polsce. Distribution Atlas of Vascular Plants in Poland.
Kraków: Laboratory of Computer Chorology - Institute of Botany - Jagiellonian University.
}
