% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atpol_grids.R
\name{.atpol10k_full}
\alias{.atpol10k_full}
\title{.atpol10k_full reads ATPOL 10km x 10km grid from file and returns it as sf object}
\usage{
.atpol10k_full()
}
\value{
Simple Feature (sf) grid of polygons for 10kmx10km ATPOL grid
}
\description{
.atpol10k_full reads ATPOL 10km x 10km grid from file and returns it as sf object
}
