% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points_on_atpol.R
\name{.atpolBackground}
\alias{.atpolBackground}
\title{.atpolBackground prepares and plots the map of ATPOL big grid (100km x 100km) on rasterized Poland}
\usage{
.atpolBackground(main = "", colors = .myCols)
}
\arguments{
\item{main}{image title, usually a species name}

\item{colors}{vector of colors to be used as a background, default internal .myCols}
}
\description{
.atpolBackground prepares and plots the map of ATPOL big grid (100km x 100km) on rasterized Poland
}
