% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data_from_old_atpol.R
\name{extract_data_from_old_atpol}
\alias{extract_data_from_old_atpol}
\title{extract_data_from_old_atpol() reads the ATPOL extracts from `.LOC` file and returns a tibble}
\usage{
extract_data_from_old_atpol(filename = )
}
\arguments{
\item{filename}{name of `.LOC` file}
}
\value{
tibble with records
}
\description{
extract_data_from_old_atpol() reads the ATPOL extracts from `.LOC` file and returns a tibble
}
\examples{
extract_data_from_old_atpol(filename = system.file("extdata/0200.LOC", package = "atpolR"))
}
