% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_to_grid.R
\name{latlon_to_grid}
\alias{latlon_to_grid}
\title{latlon_to_grid(lat, lon, length) converts geographical coordinates to ATPOL grid of given length}
\usage{
latlon_to_grid(lat, lon, length)
}
\arguments{
\item{lat}{Latitude in degrees, ex. 51.123456}

\item{lon}{Longitude in degrees, ex. 17.234567}

\item{length}{Desired ATPOL grid length, which can be 2, 4, 6, 8, 10 or 12,}
}
\value{
grid, ex. BE, BE23, BE2357, etc.
}
\description{
latlon_to_grid(lat, lon, length) converts geographical coordinates to ATPOL grid of given length
}
\examples{
latlon_to_grid(51, 17, 2)
latlon_to_grid(51, 17, 6)
}
\references{
\url{https://atpol.sourceforge.io/}
}
