% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points_on_atpol.R
\name{plotPoitsOnAtpol}
\alias{plotPoitsOnAtpol}
\title{plotPoitsOnAtpol() plots the observations on ATPOL 10km x 10km grid}
\usage{
plotPoitsOnAtpol(myData, outputType, filename, main, colors)
}
\arguments{
\item{myData}{SimpleFeature data frame with point geometry, usually centroid of ATPOL grid square}

\item{outputType}{image output type, either "svg" or "png"}

\item{filename}{name of file which will be written to disc}

\item{main}{image title, usually a species name}

\item{colors}{vector of colors to be used as a background, default internal .myCols}
}
\description{
plotPoitsOnAtpol() plots the observations on ATPOL 10km x 10km grid
}
