% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points_on_atpol.R
\name{plotPoitsOnAtpol}
\alias{plotPoitsOnAtpol}
\title{plotPoitsOnAtpol() plots the observations on ATPOL 10km x 10km grid}
\usage{
plotPoitsOnAtpol(myData, outputType, filename, main, colors, cex, col, pch)
}
\arguments{
\item{myData}{SimpleFeature data frame with point geometry, usually centroid of ATPOL grid square}

\item{outputType}{image output type, either "svg" or "png"; if not specified a standard output device is used (screen)}

\item{filename}{name of the output file}

\item{main}{image title, usually a species name}

\item{colors}{vector of colors to be used as a background, default internal .myCols}

\item{cex}{size of the points, default 0.9}

\item{col}{color of the points, default black}

\item{pch}{shape of the point, default 16 - filled dot}
}
\value{
choreograph map of species distribution in Poland.
}
\description{
plotPoitsOnAtpol() plots the observations on ATPOL 10km x 10km grid
}
